requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
openpyxl
python-dotenv
psycopg2-binary 
requests 
httpx 
pytest
pytest-asyncio 

